[Refactor] split API test files
